[#2025-01] Fix linkage on domains list page
Verified domains link directly to brand settings. Otherwise to verify.
